+++ tests/OnRamp.Test/CodeGenConsoleTest.cs
@@ -18,7 +18,7 @@ namespace OnRamp.Test
         {
             var c = new CodeGenConsole();
             var r = await c.RunAsync("--help");
-            Assert.AreEqual(0, r);
+            Assert.That(r, Is.EqualTo(0));
         }
 
         [Test]
@@ -26,7 +26,7 @@ namespace OnRamp.Test
         {
             var c = new CodeGenConsole(options: SupportedOptions.IsSimulation | SupportedOptions.ExpectNoChanges);
             var r = await c.RunAsync("--help");
-            Assert.AreEqual(0, r);
+            Assert.That(r, Is.EqualTo(0));
         }
 
         [Test]
@@ -34,7 +34,7 @@ namespace OnRamp.Test
         {
             var c = new CodeGenConsole();
             var r = await c.RunAsync("-s");
-            Assert.AreEqual(1, r);
+            Assert.That(r, Is.EqualTo(1));
         }
 
         [Test]
@@ -42,7 +42,7 @@ namespace OnRamp.Test
         {
             var c = new CodeGenConsole();
             var r = await c.RunAsync("-c");
-            Assert.AreEqual(1, r);
+            Assert.That(r, Is.EqualTo(1));
         }
 
         [Test]
@@ -50,7 +50,7 @@ namespace OnRamp.Test
         {
             var c = new CodeGenConsole();
             var r = await c.RunAsync("-d ../../Bad");
-            Assert.AreEqual(1, r);
+            Assert.That(r, Is.EqualTo(1));
         }
 
         [Test]
@@ -58,7 +58,7 @@ namespace OnRamp.Test
         {
             var c = new CodeGenConsole();
             var r = await c.RunAsync("-a NotExists");
-            Assert.AreEqual(1, r);
+            Assert.That(r, Is.EqualTo(1));
         }
 
         [Test]
@@ -66,7 +66,7 @@ namespace OnRamp.Test
         {
             var c = new CodeGenConsole();
             var r = await c.RunAsync("-db");
-            Assert.AreEqual(1, r);
+            Assert.That(r, Is.EqualTo(1));
         }
 
         [Test]
@@ -74,7 +74,7 @@ namespace OnRamp.Test
         {
             var c = new CodeGenConsole(options: SupportedOptions.IsSimulation | SupportedOptions.ExpectNoChanges);
             var r = await c.RunAsync("-c ValidEntity.yaml");
-            Assert.AreEqual(1, r);
+            Assert.That(r, Is.EqualTo(1));
         }
 
         [Test]
@@ -83,7 +83,7 @@ namespace OnRamp.Test
             var a = CodeGeneratorArgs.Create<CodeGeneratorTest>("ValidEntity.yaml");
             var c = new CodeGenConsole(a);
             var r = await c.RunAsync();
-            Assert.AreEqual(2, r);
+            Assert.That(r, Is.EqualTo(2));
         }
 
         [Test]
@@ -95,22 +95,22 @@ namespace OnRamp.Test
             var a = CodeGeneratorArgs.Create<CodeGeneratorTest>("ValidEntity.yaml", "Data/ValidEntity.yaml").AddParameter("Directory", "XA300").AddParameter("AppName", "Zzz");
             var c = new CodeGenConsole(a);
             var r = await c.RunAsync();
-            Assert.AreEqual(0, r);
+            Assert.That(r, Is.EqualTo(0));
 
-            Assert.IsTrue(Directory.Exists("XA300"));
-            Assert.AreEqual(4, Directory.GetFiles("XA300").Length);
+            Assert.That(Directory.Exists("XA300"), Is.True);
+            Assert.That(Directory.GetFiles("XA300").Length, Is.EqualTo(4));
 
-            Assert.IsTrue(File.Exists("XA300/Person.txt"));
-            Assert.AreEqual("Name: Person, CompanyName: Xxx, AppName: Zzz, Properties: Name, Age, Salary", File.ReadAllText("XA300/Person.txt"));
+            Assert.That(File.Exists("XA300/Person.txt"), Is.True);
+            Assert.That(File.ReadAllText("XA300/Person.txt"), Is.EqualTo("Name: Person, CompanyName: Xxx, AppName: Zzz, Properties: Name, Age, Salary"));
 
-            Assert.IsTrue(File.Exists("XA300/Name.txt"));
-            Assert.AreEqual("Name: Person.Name, Type: string", File.ReadAllText("XA300/Name.txt"));
+            Assert.That(File.Exists("XA300/Name.txt"), Is.True);
+            Assert.That(File.ReadAllText("XA300/Name.txt"), Is.EqualTo("Name: Person.Name, Type: string"));
 
-            Assert.IsTrue(File.Exists("XA300/Age.txt"));
-            Assert.AreEqual("Name: Person.Age, Type: int", File.ReadAllText("XA300/Age.txt"));
+            Assert.That(File.Exists("XA300/Age.txt"), Is.True);
+            Assert.That(File.ReadAllText("XA300/Age.txt"), Is.EqualTo("Name: Person.Age, Type: int"));
 
-            Assert.IsTrue(File.Exists("XA300/Salary.txt"));
-            Assert.AreEqual("Name: Person.Salary, Type: decimal?", File.ReadAllText("XA300/Salary.txt"));
+            Assert.That(File.Exists("XA300/Salary.txt"), Is.True);
+            Assert.That(File.ReadAllText("XA300/Salary.txt"), Is.EqualTo("Name: Person.Salary, Type: decimal?"));
         }
 
         [Test]
@@ -121,22 +121,22 @@ namespace OnRamp.Test
 
             var c = new CodeGenConsole();
             var r = await c.RunAsync("-s ValidEntity.yaml -c Data/ValidEntity.yaml -p Directory=XA310 -p AppName=Zzz -a \"OnRamp.Test, Version=1.2.3.0, Culture=neutral, PublicKeyToken=null\"");
-            Assert.AreEqual(0, r);
+            Assert.That(r, Is.EqualTo(0));
 
-            Assert.IsTrue(Directory.Exists("XA310"));
-            Assert.AreEqual(4, Directory.GetFiles("XA310").Length);
+            Assert.That(Directory.Exists("XA310"), Is.True);
+            Assert.That(Directory.GetFiles("XA310").Length, Is.EqualTo(4));
 
-            Assert.IsTrue(File.Exists("XA310/Person.txt"));
-            Assert.AreEqual("Name: Person, CompanyName: Xxx, AppName: Zzz, Properties: Name, Age, Salary", File.ReadAllText("XA310/Person.txt"));
+            Assert.That(File.Exists("XA310/Person.txt"), Is.True);
+            Assert.That(File.ReadAllText("XA310/Person.txt"), Is.EqualTo("Name: Person, CompanyName: Xxx, AppName: Zzz, Properties: Name, Age, Salary"));
 
-            Assert.IsTrue(File.Exists("XA310/Name.txt"));
-            Assert.AreEqual("Name: Person.Name, Type: string", File.ReadAllText("XA310/Name.txt"));
+            Assert.That(File.Exists("XA310/Name.txt"), Is.True);
+            Assert.That(File.ReadAllText("XA310/Name.txt"), Is.EqualTo("Name: Person.Name, Type: string"));
 
-            Assert.IsTrue(File.Exists("XA310/Age.txt"));
-            Assert.AreEqual("Name: Person.Age, Type: int", File.ReadAllText("XA310/Age.txt"));
+            Assert.That(File.Exists("XA310/Age.txt"), Is.True);
+            Assert.That(File.ReadAllText("XA310/Age.txt"), Is.EqualTo("Name: Person.Age, Type: int"));
 
-            Assert.IsTrue(File.Exists("XA310/Salary.txt"));
-            Assert.AreEqual("Name: Person.Salary, Type: decimal?", File.ReadAllText("XA310/Salary.txt"));
+            Assert.That(File.Exists("XA310/Salary.txt"), Is.True);
+            Assert.That(File.ReadAllText("XA310/Salary.txt"), Is.EqualTo("Name: Person.Salary, Type: decimal?"));
         }
 
         [Test]
@@ -148,33 +148,33 @@ namespace OnRamp.Test
             // Run and it should fail as it cannot create.
             var c = new CodeGenConsole();
             var r = await c.RunAsync("-s ValidEntity.yaml -c Data/ValidEntity.yaml -enc -p Directory=XA400 -p AppName=Zzz -a \"OnRamp.Test, Version=1.2.3.0, Culture=neutral, PublicKeyToken=null\"");
-            Assert.AreEqual(3, r);
+            Assert.That(r, Is.EqualTo(3));
 
             // Run again and let it make changes.
             c = new CodeGenConsole();
             r = await c.RunAsync("-s ValidEntity.yaml -c Data/ValidEntity.yaml -p Directory=XA400 -p AppName=Zzz -a \"OnRamp.Test, Version=1.2.3.0, Culture=neutral, PublicKeyToken=null\"");
-            Assert.AreEqual(0, r);
+            Assert.That(r, Is.EqualTo(0));
 
             // Change the file and run again and it should fail as it cannot update.
             var files = Directory.GetFiles("XA400");
-            Assert.AreNotEqual(0, files.Length);
+            Assert.That(files.Length, Is.Not.EqualTo(0));
 
             var content = File.ReadAllText(files.Last());
             File.WriteAllText(files.Last(), content + "X");
 
             c = new CodeGenConsole();
             r = await c.RunAsync("-s ValidEntity.yaml -c Data/ValidEntity.yaml -enc -p Directory=XA400 -p AppName=Zzz -a \"OnRamp.Test, Version=1.2.3.0, Culture=neutral, PublicKeyToken=null\"");
-            Assert.AreEqual(3, r);
+            Assert.That(r, Is.EqualTo(3));
         }
 
         [Test]
         public void C100_GetBaseExeDirectory()
         {
             var ed = Environment.CurrentDirectory;
-            Assert.IsTrue(ed.Contains(Path.Combine("bin", "debug"), StringComparison.InvariantCultureIgnoreCase) || ed.Contains(Path.Combine("bin", "release"), StringComparison.InvariantCultureIgnoreCase));
+            Assert.That(ed.Contains(Path.Combine("bin", "debug"), StringComparison.InvariantCultureIgnoreCase) || ed.Contains(Path.Combine("bin", "release"), StringComparison.InvariantCultureIgnoreCase), Is.True);
 
             ed = CodeGenConsole.GetBaseExeDirectory();
-            Assert.IsFalse(ed.Contains(Path.Combine("bin", "debug"), StringComparison.InvariantCultureIgnoreCase) || ed.Contains(Path.Combine("bin", "release"), StringComparison.InvariantCultureIgnoreCase));
+            Assert.That(ed.Contains(Path.Combine("bin", "debug"), StringComparison.InvariantCultureIgnoreCase) || ed.Contains(Path.Combine("bin", "release"), StringComparison.InvariantCultureIgnoreCase), Is.False);
         }
     }
 }