+++ tests/OnRamp.Test/Config/PropertyConfig.cs
@@ -1,27 +1,26 @@
 ﻿using OnRamp.Config;
-using Newtonsoft.Json;
+using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 
 #nullable enable
 
 namespace OnRamp.Test.Config
 {
-    [JsonObject(MemberSerialization = MemberSerialization.OptIn)]
     [CodeGenClass("Property", Title = "'Property' object.", Description = "The `Property` object.")]
     [CodeGenCategory("Key", Title = "Provides the _Key_ configuration.")]
     public class PropertyConfig : ConfigBase<EntityConfig, EntityConfig>
     {
         public override string QualifiedKeyName => BuildQualifiedKeyName("Property", Name);
 
-        [JsonProperty("name")]
+        [JsonPropertyName("name")]
         [CodeGenProperty("Key", Title = "The property name.", IsMandatory = true, IsUnique = true)]
         public string? Name { get; set; }
 
-        [JsonProperty("type")]
+        [JsonPropertyName("type")]
         [CodeGenProperty("Key", Title = "The property type.", Description = "This is a more detailed description for the property type.", IsImportant = true, Options = new string[] { "string", "int", "decimal" })]
         public string? Type { get; set; }
 
-        [JsonProperty("isNullable")]
+        [JsonPropertyName("isNullable")]
         [CodeGenProperty("Key", Title = "Indicates whether the property is nullable.")]
         public bool? IsNullable { get; set; }